+++ src/model/playerClasses/Druid.java
@@ -1,17 +1,17 @@
-package playerClasses;
+package model.playerClasses;
 
 import enums.PlayerClass;
 
-public class Ninja implements SetStats{
-    private int attack = 210;
+public class Druid implements SetStats {
+    private int attack = 70;
     private int defense = 60;
-    private int health = 70;
-    private int specialAttack = 0;
+    private int health = 100 ;
+    private int specialAttack = 10;
     private int specialDefense = 60;
 
     @Override
     public void SetPlayerStatsForClass(PlayerClass playerClass) {
-        System.out.println("The Players calss : " + playerClass);
+        System.out.println("Player Class: " + playerClass);
     }
 
     public int getAttack() {
