+++ src/model/playerClasses/Ninja.java
@@ -1,17 +1,17 @@
-package playerClasses;
+package model.playerClasses;
 
 import enums.PlayerClass;
 
-public class Paladin implements SetStats{
-    private int attack = 70;
+public class Ninja implements SetStats{
+    private int attack = 210;
     private int defense = 60;
-    private int health = 100;
-    private int specialAttack = 10;
+    private int health = 70;
+    private int specialAttack = 0;
     private int specialDefense = 60;
 
     @Override
     public void SetPlayerStatsForClass(PlayerClass playerClass) {
-        System.out.println("The players class " + playerClass);
+        System.out.println("The Players calss : " + playerClass);
     }
 
     public int getAttack() {
