+++ src/model/playerClasses/Paladin.java
@@ -1,17 +1,17 @@
-package playerClasses;
+package model.playerClasses;
 
 import enums.PlayerClass;
 
-public class Druid implements SetStats {
+public class Paladin implements SetStats{
     private int attack = 70;
     private int defense = 60;
-    private int health = 100 ;
+    private int health = 100;
     private int specialAttack = 10;
     private int specialDefense = 60;
 
     @Override
     public void SetPlayerStatsForClass(PlayerClass playerClass) {
-        System.out.println("Player Class: " + playerClass);
+        System.out.println("The players class " + playerClass);
     }
 
     public int getAttack() {
